requirements.txt
# Core MCP dependencies
mcp[cli]>=1.4.1
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
ton

# HTTP client
aiohttp>=3.8.0
httpx>=0.24.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# JSON handling
orjson>=3.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging
structlog>=23.0.0

# Date/time handling
python-dateutil>=2.8.0

# Async support
anyio>=3.7.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: for enhanced NLP processing
spacy>=3.6.0
nltk>=3.8.0